ok, better workflow

+++ src/publish.js
@@ -1,42 +1,11 @@
 // Imports
 const core = require("@actions/core");
 const shell = require("shelljs");
+const artifact = require("@actions/artifact");
 
 // Exports
 module.exports = { publishOutput };
 
-// Get repo info
-
-var TARGET_BRANCH;
-const GITHUB_TOKEN = core.getInput("github-token");
-const GITHUB_SLUG = shell.env["GITHUB_REPOSITORY"];
-const EVENT_NAME = shell.env["GITHUB_EVENT_NAME"];
-const GITHUB_WORKSPACE = shell.env["GITHUB_WORKSPACE"];
-const GITHUB_REF = shell.env["GITHUB_REF"];
-const OUTPUT_BRANCH_SUFFIX = core.getInput("output-branch-suffix");
-if (EVENT_NAME == 'pull_request') {
-
-  // This is a pull request, so we'll force-push the output
-  // to `pull-request-<number>-pdf`
-  const PULL_REQUEST_NUMBER = GITHUB_REF.split("/")[2];
-  TARGET_BRANCH = `pull-request-${PULL_REQUEST_NUMBER}-pdf`;
-
-} else {
-
-  // Not a pull request, so we'll force-push the output
-  // to `<branch_name>-pdf`
-  const GITHUB_BRANCH = GITHUB_REF.split("/")[2];
-  TARGET_BRANCH = `${GITHUB_BRANCH}-${OUTPUT_BRANCH_SUFFIX}`;
-  
-
-}
-
-// DEBUG DEBUG DEBUG
-shell.echo("DEBUG");
-shell.echo(GITHUB_TOKEN[0]);
-shell.echo(GITHUB_TOKEN[5]);
-shell.echo(GITHUB_TOKEN[10]);
-shell.echo("DEBUG");
 
 /**
  * Publish the article output.
@@ -48,35 +17,52 @@ async function publishOutput() {
   const config = require(`${GITHUB_WORKSPACE}/.showyourwork/config.json`);
   const output = [config["ms_pdf"]];
 
-  // Upload the arxiv tarball?
+  // Include the arxiv tarball?
   if (core.getInput("build-tarball") == "true") {
     output.push("arxiv.tar.gz");
   }
 
-  // Force-push output to a separate branch
-  core.startGroup("Uploading output");
-  const TARGET_DIRECTORY = shell
-    .exec("mktemp -d")
-    .replace(/(\r\n|\n|\r)/gm, "");
-  shell.cp("-R", ".", `${TARGET_DIRECTORY}`);
-  shell.cd(`${TARGET_DIRECTORY}`);
-  shell.exec(`git checkout --orphan ${TARGET_BRANCH}`);
-  shell.exec("git rm --cached -rf .", {silent: true});
-  for (const out of output) {
-    shell.exec(`git add -f ${out}`);
-  }
-  shell.exec(
-    "git -c user.name='showyourwork' -c user.email='showyourwork' " +
-      "commit -m 'force-push article output'"
+  // Upload an artifact
+  const artifactClient = artifact.create();
+  const uploadResponse = await artifactClient.uploadArtifact(
+    "showyourwork-output", 
+    output, 
+    ".", 
+    {
+      continueOnError: false
+    }
   );
-  shell.exec(
-    "git push --force " +
-      `https://x-access-token:${GITHUB_TOKEN}@github.com/${GITHUB_SLUG} ` +
-      `${TARGET_BRANCH}`
-  );
-  shell.cd(GITHUB_WORKSPACE);
-  core.endGroup();
 
-  // Set an action output containing the link to the PDF
-  core.setOutput("pdf-url", `https://github.com/${GITHUB_SLUG}/raw/${TARGET_BRANCH}/${config["ms_pdf"]}`);
+  // Force-push output to a separate branch (if not a PR)
+  if (shell.env["GITHUB_EVENT_NAME"] != 'pull_request') {
+    core.startGroup("Uploading output");
+    const GITHUB_TOKEN = core.getInput("github-token");
+    const GITHUB_SLUG = shell.env["GITHUB_REPOSITORY"];
+    const GITHUB_WORKSPACE = shell.env["GITHUB_WORKSPACE"];
+    const GITHUB_REF = shell.env["GITHUB_REF"];
+    const OUTPUT_BRANCH_SUFFIX = core.getInput("output-branch-suffix");
+    const GITHUB_BRANCH = GITHUB_REF.split("/")[2];
+    const TARGET_BRANCH = `${GITHUB_BRANCH}-${OUTPUT_BRANCH_SUFFIX}`;
+    const TARGET_DIRECTORY = shell
+      .exec("mktemp -d")
+      .replace(/(\r\n|\n|\r)/gm, "");
+    shell.cp("-R", ".", `${TARGET_DIRECTORY}`);
+    shell.cd(`${TARGET_DIRECTORY}`);
+    shell.exec(`git checkout --orphan ${TARGET_BRANCH}`);
+    shell.exec("git rm --cached -rf .", {silent: true});
+    for (const out of output) {
+      shell.exec(`git add -f ${out}`);
+    }
+    shell.exec(
+      "git -c user.name='showyourwork' -c user.email='showyourwork' " +
+        "commit -m 'force-push article output'"
+    );
+    shell.exec(
+      "git push --force " +
+        `https://x-access-token:${GITHUB_TOKEN}@github.com/${GITHUB_SLUG} ` +
+        `${TARGET_BRANCH}`
+    );
+    shell.cd(GITHUB_WORKSPACE);
+    core.endGroup();
+  }
 }
